+++ chemreac/serialization.py
@@ -42,8 +42,8 @@ def dump(rd, dest):
         'D': list(rd._D),
         'mobility': list(rd._mobility),
         # 'x': rd.x.tolist(),
-        'stoich_inactv': rd.stoich_inactv,
-        'unit_registry': unit_registry_to_human_readable(rd.unit_registry),
+        'stoich_inact': rd.stoich_inact,
+        'units': unit_registry_to_human_readable(rd.unit_registry),
         'g_values': list(rd._g_values),
         'g_value_parents': rd.g_value_parents,
         # 'fields': fields
@@ -73,8 +73,9 @@ def load(source, RD=None, **kwargs):
         fh = source
 
     data = json.load(fh)
-    unit_registry = data['unit_registry'] = unit_registry_from_human_readable(
-        data.get('unit_registry', None))
+    units = data.pop('units', None)
+    unit_registry = unit_registry_from_human_readable(units)
+    data['unit_registry'] = unit_registry
     if 'D' in data:
         data['D'] = np.array(data['D'])*get_unit(unit_registry, 'diffusion')
     if 'mobility' in data:
@@ -82,8 +83,9 @@ def load(source, RD=None, **kwargs):
             unit_registry, 'electrical_mobility')
     if 'g_values' in data:
         data['g_values'] = [elem*g_unit for elem, g_unit in
-                            zip(data['g_values'], RD.g_units(
-                                unit_registry, data['g_value_parents']))]
+                            zip(np.array(data['g_values']),
+                                RD.g_units(unit_registry,
+                                           data['g_value_parents']))]
     reaction_orders = map(len, data['stoich_active'])
     kunits = RD.k_units(unit_registry, reaction_orders)
     data['k'] = [kval*kunit for kval, kunit in zip(data['k'], kunits)]