+++ examples/four_species.py
@@ -47,11 +47,11 @@ import os
 import argh
 import numpy as np
 
+from chempy.util.graph import rsys2graph
 from chemreac import DENSE, BANDED
 from chemreac.chemistry import ReactionSystem, mk_sn_dict_from_names
 from chemreac.integrate import run
 from chemreac.serialization import load
-from chemreac.util.graph import rsys2graph
 from chemreac.util.plotting import (
     coloured_spy, plot_jacobian, plot_per_reaction_contribution,
     save_and_or_show_plot